Modify for Newton

+++ run_test.sh
@@ -3,12 +3,26 @@
 export CINDER_DIR=./cinder
 export CINDER_REPO_URL="https://git.openstack.org/openstack/cinder"
 
-echo "${CINDER_DIR}"
-if [ -d "${CINDER_DIR}" ]; then
-    rm -rf $CINDER_DIR
+echo "Running Flake8..."
+flake8 infortrend/
+if [ $? -ne 0 ]; then
+    exit 1
 fi
+flake8 test/infortrend/
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+echo "Complete."
 
-git clone $CINDER_REPO_URL --depth=1 -b stable/newton
+if [ -z "${2}" ]; then
+    if [ -d "${CINDER_DIR}" ]; then
+        echo "Deleting $CINDER_DIR"
+        rm -rf $CINDER_DIR
+    fi
+    git clone $CINDER_REPO_URL -b driverfixes/newton --depth=1
+else
+    echo "Skip Cloning cinder."
+fi
 
 source setupIFTdriver.sh $CINDER_DIR
 
@@ -21,6 +35,6 @@ else
     sed -i "s#stestr run '{posargs}'#stestr run '{posargs}' --concurrency=4#g" tox.ini
 fi
 
-tox -e ${1} test_infortrend_
+tox -e ${1} test_infortrend_ -- --concurrency=8
 
 cd ..
